Create ViewModel
