+++ GuessTheFlag/GuessTheFlagApp.swift
@@ -14,7 +14,7 @@ import SwiftUI
 struct GuessTheFlagApp: App {
     var body: some Scene {
         WindowGroup {
-            ContentView()
+            GameView()
         }
     }
 }
